Wall Jump bug fixes
Fixed bug where not letting go of input causes the Z velocity to remain clamped, preventing the character from wall jumping a second time in certain scenarios. The only bug remaining after this is I don't want the character to be able to scale a single wall by hugging it.