Cage Trap update!
Created a hanging cage trap that falls down on the unsuspecting player. If the player is hit by the falling cage (after it hits the ground), they will be teleported to its center and prevented from escaping. This is only halfway complete. I need to create animations for breaking the cage open and give it a health system and hit reaction particles.